Refactor, move main and filters to separate component
